requirements.txt
ass==0.5.4
bcp47==0.1.0
beautifulsoup4==4.12.3
bitmath==1.3.3.1
bs4==0.0.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
deep-translator==1.11.4
ffmpeg-python==0.2.0
future==1.0.0
idna==3.10
iso-639==0.4.5
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
numpy==2.1.2
pydub==0.25.1
pymkv==1.0.8
pysrt==1.1.2
requests==2.32.3
setuptools==75.2.0
soupsieve==2.6
urllib3==2.2.3
